Moved relevant files into utils directory
Data files and colour converter moved to utils
