+++ src/utils/emotionScores.js
